// apps/web-svelte/src/lib/api/generated/model/acquisitionTransferView.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface AcquisitionTransferView {
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$ */
  progress: number | string;
  /** @nullable */
  state: string | null;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  totalSizeBytes: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$ */
  downloadSpeedBytesPerSecond: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  etaSeconds: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  seeds: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  peers: number | string;
  /** @nullable */
  savePath: string | null;
  pieceStates: (number | string)[];
}